podvm: remove cdh,attestation-agent units
As we build kata with SEALED_SECRET=yes by default this implies that
kata-agent will attempt to spawn attestation-agent, cdh and
api-server-rest. We'll end up with duplicate processes and contention
over the sockets they need to create.

We can/need to keep api-server-rest as a systemd unit since this one
needs to run in the podns network namespace and since it's exposing a
tcp socket there is no contention.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>