e2e-test: assert job condition on job tests
We have elaborate condition checks on a job's pods, but we do not test
the job status itself, which will produce wrong test results. A test
for job that always fails (cmd: /bin/false) will pass.

This keeps the current assertions in-place, but makes the test fail on a
failed job status

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>