ci: fix azure nightly mkosi build
Due to a rebase fail, the workflow attempted to use the ubuntu mkosi
build. We default to build using nix to stay close to the documented
procedure. If we opt to use the ubuntu-packaged mkosi, more testing is
required and we should change it everywhere.

In the mid-term this makes sense, because currently mkosi is marked as
broken in nixos-24.05 and nix-unstable and the version we have pinned
(v17) is rather old now. Later versions of mkosi also don't require root
privileges anymore.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>