podvm: add scratch-space logic
This adds the configuration for an encrypted scratch space in /dev/sda4
on an mkosi image. If the image has space available it will create a
"scratch" partition and encrypt it using an adhoc LUKS key.

An empty `/run/peerpod/mount-scratch` file is being added to the
write-files directive of userdata, if a disk size param has been set. In
this case kata-agent will mount the encrypted scratch space to
`/run/kata-containers` prior to startup.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>